Setup max tokens dialog
